// prismicio-types.d.ts
type Simplify<T> = { [KeyType in keyof T]: T[KeyType] };
// ...
export interface AssessmentGoalsPageDocumentDataGoalOverviewLabelsItem {
  /**
   * Done field in *Assessment goals page → Goal overview labels*
   *
   * - **Field Type**: Text
   * - **Placeholder**: Done
   * - **API ID Path**: assessment_goals_page.goal_overview_labels[].done
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  done: prismic.KeyTextField;

  /**
   * Edit field in *Assessment goals page → Goal overview labels*
   *
   * - **Field Type**: Text
   * - **Placeholder**: edit
   * - **API ID Path**: assessment_goals_page.goal_overview_labels[].edit
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  edit: prismic.KeyTextField;
}

/**
 * Item in *Assessment goals page → Goal priorities*
 */
export interface AssessmentGoalsPageDocumentDataGoalPrioritiesItem {
  /**
   * label field in *Assessment goals page → Goal priorities*
   *
   * - **Field Type**: Text
   * - **Placeholder**: Relevant/Not Relevant/Don't know
   * - **API ID Path**: assessment_goals_page.goal_priorities[].label
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  label: prismic.KeyTextField;

  /**
   * value field in *Assessment goals page → Goal priorities*
   *
   * - **Field Type**: Number
   * - **Placeholder**: Set to 1 for "Relevant", 2 for "Not relevant", 3 for "Don't know"
   * - **API ID Path**: assessment_goals_page.goal_priorities[].value
   * - **Documentation**: https://prismic.io/docs/field#number
   */
  value: prismic.NumberField;
}

/**
 * Item in *Assessment goals page → Summary*
 */
export interface AssessmentGoalsPageDocumentDataSummaryItem {
  /**
   * Summary description field in *Assessment goals page → Summary*
   *
   * - **Field Type**: Text
   * - **Placeholder**: This is the assessment board. Click on each goal and complete all assessments before you can see your result.
   * - **API ID Path**: assessment_goals_page.summary[].summary_description
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  summary_description: prismic.KeyTextField;

  /**
   * Button label field in *Assessment goals page → Summary*
   *
   * - **Field Type**: Text
   * - **Placeholder**: See result
   * - **API ID Path**: assessment_goals_page.summary[].button_label
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  button_label: prismic.KeyTextField;

  /**
   * Count suffix field in *Assessment goals page → Summary*
   *
   * - **Field Type**: Text
   * - **Placeholder**: Label for count suffix ("goals completed")
   * - **API ID Path**: assessment_goals_page.summary[].count_suffix
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  count_suffix: prismic.KeyTextField;
}

/**
 * Item in *Assessment goals page → Goal sorting modal*
 */
export interface AssessmentGoalsPageDocumentDataGoalSortingModalItem {
  /**
   * Title field in *Assessment goals page → Goal sorting modal*
   *
   * - **Field Type**: Text
   * - **Placeholder**: How to do the SDG impact assessment
   * - **API ID Path**: assessment_goals_page.goal_sorting_modal[].title
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  title: prismic.KeyTextField;

  /**
   * Sorting instructions field in *Assessment goals page → Goal sorting modal*
   *
   * - **Field Type**: Rich Text
   * - **Placeholder**: Instuctions for sorting process
   * - **API ID Path**: assessment_goals_page.goal_sorting_modal[].sorting_instructions
   * - **Documentation**: https://prismic.io/docs/field#rich-text-title
   */
  sorting_instructions: prismic.RichTextField;

  /**
   * Button label field in *Assessment goals page → Goal sorting modal*
   *
   * - **Field Type**: Text
   * - **Placeholder**: Start sorting
   * - **API ID Path**: assessment_goals_page.goal_sorting_modal[].button_label
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  button_label: prismic.KeyTextField;
}

/**
 * Content for Assessment goals page documents
 */
interface AssessmentGoalsPageDocumentData {
  /**
   * description field in *Assessment goals page*
   *
   * - **Field Type**: Rich Text
   * - **Placeholder**: Perform the impact assessment by clicking on any SDG
   * - **API ID Path**: assessment_goals_page.description
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#rich-text-title
   */
  description: prismic.RichTextField;

  /**
   * Goal overview labels field in *Assessment goals page*
   *
   * - **Field Type**: Group
   * - **Placeholder**: *None*
   * - **API ID Path**: assessment_goals_page.goal_overview_labels[]
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#group
   */
  goal_overview_labels: prismic.GroupField<
    Simplify<AssessmentGoalsPageDocumentDataGoalOverviewLabelsItem>
  >;

  /**
   * Save and close field in *Assessment goals page*
   *
   * - **Field Type**: Text
   * - **Placeholder**: Save & close
   * - **API ID Path**: assessment_goals_page.save_and_close
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  save_and_close: prismic.KeyTextField;

  /**
   * Goal priorities field in *Assessment goals page*
   *
   * - **Field Type**: Group
   * - **Placeholder**: *None*
   * - **API ID Path**: assessment_goals_page.goal_priorities[]
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#group
   */
  goal_priorities: prismic.GroupField<
    Simplify<AssessmentGoalsPageDocumentDataGoalPrioritiesItem>
  >;

  /**
   * Summary field in *Assessment goals page*
   *
   * - **Field Type**: Group
   * - **Placeholder**: *None*
   * - **API ID Path**: assessment_goals_page.summary[]
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#group
   */
  summary: prismic.GroupField<
    Simplify<AssessmentGoalsPageDocumentDataSummaryItem>
  >;

  /**
   * Goal sorting modal field in *Assessment goals page*
   *
   * - **Field Type**: Group
   * - **Placeholder**: *None*
   * - **API ID Path**: assessment_goals_page.goal_sorting_modal[]
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#group
   */
  goal_sorting_modal: prismic.GroupField<
    Simplify<AssessmentGoalsPageDocumentDataGoalSortingModalItem>
  >;
}

/**
 * Assessment goals page document from Prismic
 *
 * - **API ID**: `assessment_goals_page`
 * - **Repeatable**: `false`
 * - **Documentation**: https://prismic.io/docs/custom-types
 *
 * @typeParam Lang - Language API ID of the document.
 */
export type AssessmentGoalsPageDocument<Lang extends string = string> =
  prismic.PrismicDocumentWithoutUID<
    Simplify<AssessmentGoalsPageDocumentData>,
    "assessment_goals_page",
    Lang
  >;
// ...
type GoalsDocumentDataBodySlice = GoalsDocumentDataBodyGoalSlice;

/**
 * Content for Goals documents
 */
interface GoalsDocumentData {
  /**
   * Slice zone field in *Goals*
   *
   * - **Field Type**: Slice Zone
   * - **Placeholder**: *None*
   * - **API ID Path**: goals.body[]
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#slices
   */
  body: prismic.SliceZone<GoalsDocumentDataBodySlice>;
}

/**
 * Goals document from Prismic
 *
 * - **API ID**: `goals`
 * - **Repeatable**: `false`
 * - **Documentation**: https://prismic.io/docs/custom-types
 *
 * @typeParam Lang - Language API ID of the document.
 */
export type GoalsDocument<Lang extends string = string> =
  prismic.PrismicDocumentWithUID<Simplify<GoalsDocumentData>, "goals", Lang>;